+++ spec/simple.spec.ts
@@ -64,42 +64,6 @@ describe("late", () => {
   });
 });
 
-describe("maybe", () => {
-  const maybeType = types.maybe(types.string);
-
-  test("can create a read-only instance", () => {
-    expect(maybeType.createReadOnly()).toEqual(undefined);
-    expect(maybeType.createReadOnly(undefined)).toEqual(undefined);
-    expect(maybeType.createReadOnly("value 2")).toEqual("value 2");
-  });
-
-  test("can be verified with is", () => {
-    expect(maybeType.is(undefined)).toEqual(true);
-    expect(maybeType.is("not testing")).toEqual(true);
-    expect(maybeType.is(null)).toEqual(false);
-    expect(maybeType.is(true)).toEqual(false);
-    expect(maybeType.is({})).toEqual(false);
-  });
-});
-
-describe("maybeNull", () => {
-  const maybeNullType = types.maybeNull(types.string);
-
-  test("can create a read-only instance", () => {
-    expect(maybeNullType.createReadOnly(null)).toEqual(null);
-    expect(maybeNullType.createReadOnly(undefined)).toEqual(null);
-    expect(maybeNullType.createReadOnly("value 2")).toEqual("value 2");
-  });
-
-  test("can be verified with is", () => {
-    expect(maybeNullType.is(null)).toEqual(true);
-    expect(maybeNullType.is(undefined)).toEqual(true);
-    expect(maybeNullType.is("not testing")).toEqual(true);
-    expect(maybeNullType.is(true)).toEqual(false);
-    expect(maybeNullType.is({})).toEqual(false);
-  });
-});
-
 describe("literal", () => {
   const literal = types.literal("testing");
